Combat Integration
Added a new scene that is identical to Johnny's central area but with
main character combat and a rat. Both the fps controller and the rat
have a prefab inside of jesse's assets/_prefabs. Some public variables
in scripts need to be set in the inspector before they will function
properly though. Most needs are self explanatory, but if they aren't
just ask me and I can help explain what it wants.